Refactor into separate link room feature
