+++ app/lib/common/widgets/room/room_hierarchy_options_menu.dart
@@ -1,6 +1,6 @@
 import 'package:acter/common/providers/room_providers.dart';
 import 'package:acter/features/space/actions/set_child_room_suggested.dart';
-import 'package:acter/features/space/actions/unlink_child_room.dart';
+import 'package:acter/features/link_room/actions/unlink_child_room.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
